Desktop/DigiSkola-DE/project 1 final/data-engineer/Digital Skola/Project/Project 1/requirements.txt
pandas
python-dotenv
SQLAlchemy